feat: add account deletion form

+++ frontend/src/pages/ProfilePage/ProfilePage.tsx
@@ -7,8 +7,8 @@ import {
   signOut,
 } from "supertokens-auth-react/recipe/session"
 import { toast } from "sonner"
-import { MdDelete } from "react-icons/md"
 import { deleteAccount, getUserEmail } from "../../api/auth/account.ts"
+import ProfileDeleteForm from "../../features/profile/delete/ProfileDeleteForm/ProfileDeleteForm.tsx"
 import LoadingDisplay from "../../components/LoadingDisplay/LoadingDisplay.tsx"
 import Button from "../../components/ui/button/Button.tsx"
 import Separator from "../../components/Separator/Separator.tsx"
@@ -50,35 +50,42 @@ function ProfilePage() {
     })
   }, [session, navigateToHomepage])
 
-  const handleDeleteAccount = useCallback(async () => {
-    if (session.loading) {
-      return
-    }
-    const sessionExists = await doesSessionExist()
-    if (!sessionExists) {
-      toast.info("Session expired. Please login again")
-      await signOut()
-      navigateToHomepage()
-      return
-    }
-    if (!session.userId) {
-      return
-    }
-    try {
-      setLoading(true)
-      await deleteAccount(
-        session.accessTokenPayload.supabase_token,
-        session.userId
-      )
-      toast.success("Account deleted")
-      navigateToHomepage()
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (error: any) {
-      toast.error(error.message)
-    } finally {
-      setLoading(false)
-    }
-  }, [session, navigateToHomepage])
+  const handleDeleteAccount = useCallback(
+    async (email: string) => {
+      if (session.loading) {
+        return
+      }
+      const sessionExists = await doesSessionExist()
+      if (!sessionExists) {
+        toast.info("Session expired. Please login again")
+        await signOut()
+        navigateToHomepage()
+        return
+      }
+      if (!session.userId || userEmail === "") {
+        return
+      }
+      if (email !== userEmail) {
+        toast.error("Could not delete account. Provided email does not match")
+        return
+      }
+      try {
+        setLoading(true)
+        await deleteAccount(
+          session.accessTokenPayload.supabase_token,
+          session.userId
+        )
+        toast.success("Account deleted")
+        navigateToHomepage()
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      } catch (error: any) {
+        toast.error(error.message)
+      } finally {
+        setLoading(false)
+      }
+    },
+    [userEmail, session, navigateToHomepage]
+  )
 
   const handleLogoutAccount = useCallback(async () => {
     setLoading(true)
@@ -93,13 +100,15 @@ function ProfilePage() {
   }, [navigate])
 
   return (
-    <div className="profile-page-container">
-      <h2>Profile</h2>
-      <Separator />
-      <LoadingDisplay loading={loading}>
-        <p>
-          <b>Email:</b> {userEmail}
-        </p>
+    <LoadingDisplay loading={loading}>
+      <div className="profile-page-container">
+        <div>
+          <h2>Profile</h2>
+          <Separator />
+          <p className="profile-page-email">
+            <b>Email:</b> {userEmail}
+          </p>
+        </div>
         <div className="profile-page-account-actions-container">
           <Button
             keyProp="profile-page-reset-password-button"
@@ -109,27 +118,18 @@ function ProfilePage() {
           >
             Reset Password
           </Button>
+          <Button
+            keyProp="profile-page-logout-button"
+            onClick={handleLogoutAccount}
+            variant="primary"
+            title="Logout"
+          >
+            Logout
+          </Button>
         </div>
-        <Button
-          keyProp="profile-page-logout-button"
-          onClick={handleLogoutAccount}
-          variant="primary"
-          title="Logout"
-        >
-          Logout
-        </Button>
-        <Button
-          keyProp="profile-page-delete-profile-button"
-          onClick={handleDeleteAccount}
-          className="delete-profile-button"
-          variant="danger"
-          title="Delete Account"
-        >
-          <MdDelete size={24} />
-          Delete Account
-        </Button>
-      </LoadingDisplay>
-    </div>
+        <ProfileDeleteForm onDelete={handleDeleteAccount} />
+      </div>
+    </LoadingDisplay>
   )
 }
